+++ libft/srcs/math-functions/ft_count_digits.c
@@ -12,7 +12,7 @@
 
 #include "../../includes/libft.h"
 
-int		ft_count_digits(long long n)
+int		ft_count_digits(LL n)
 {
 	int digits;
